requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
browser-cookie3==0.20.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chromedriver-autoinstaller==0.6.4
click==8.1.8
colorama==0.4.6
distlib==0.3.9
filelock==3.17.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2025.3.0
h11==0.14.0
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lz4==4.4.3
MarkupSafe==3.0.2
multidict==6.1.0
mysql-connector==2.2.9
mysql-connector-python==9.2.0
numpy==2.2.4
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pipenv==2024.4.1
platformdirs==4.3.6
pool==0.1.2.dev0
propcache==0.3.0
pycparser==2.22
pycryptodomex==3.22.0
PySocks==1.7.1
pytesseract==0.3.13
python-dotenv==1.0.1
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
selenium==4.28.1
setuptools==75.8.0
shadowcopy==0.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
virtualenv==20.29.2
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
WMI==1.5.1
wsproto==1.2.0
yarl==1.18.3

